requirements.txt
GitPython
aicli
PyYAML>=6.0.1
requests
setuptools-rust
wheel
paramiko==2.12.0
tenacity
gspread
oauth2client
jinja2
kubernetes
python-hosts
argcomplete
yq
pyserial
